Moved the refresh stuff into startup, so it isn't messing up the class defs.
